+++ static/js/server_events_dispatch.js
@@ -597,6 +597,7 @@ export function dispatch_normal_event(event) {
                 "fluid_layout_width",
                 "high_contrast_mode",
                 "left_side_userlist",
+                "realm_icon_as_favicon",
                 "timezone",
                 "twenty_four_hour_time",
                 "translate_emoticons",
@@ -635,6 +636,9 @@ export function dispatch_normal_event(event) {
             if (event.property === "dense_mode") {
                 $("body").toggleClass("less_dense_mode");
                 $("body").toggleClass("more_dense_mode");
+            }
+            if (event.property === "realm_icon_as_favicon") {
+                notifications.redraw_favicon();
             }
             if (event.property === "color_scheme") {
                 $("body").fadeOut(300);